+++ projects/igniteui-angular/src/lib/grids/common/events.ts
@@ -8,7 +8,6 @@ import { IgxColumnComponent } from '../columns/column.component';
 import { IgxGridBaseDirective } from '../grid-base.directive';
 import { IgxRowDirective } from '../row.directive';
 import { ColumnType } from './column.interface';
-import { IgxCell } from '../selection/selection.service';
 export { GridSelectionRange } from '../selection/selection.service';
 
 export interface IGridClipboardEvent {
@@ -28,11 +27,12 @@ export interface IGridEditEventArgs extends CancelableEventArgs, IBaseEventArgs 
         columnID: any,
         rowIndex: number
     };
+    rowData: any;
     oldValue: any;
     newValue?: any;
     event?: Event;
-    cell?: IgxCell;
     column?: ColumnType;
+    owner?: IgxGridBaseDirective & GridType;
 }
 
 export interface IPinColumnEventArgs extends IBaseEventArgs {