+++ projects/igniteui-angular/src/lib/grids/grid/grid.crud.spec.ts
@@ -7,7 +7,6 @@ import { IgxGridModule } from './public_api';
 import { wait } from '../../test-utils/ui-interactions.spec';
 import { configureTestSuite } from '../../test-utils/configure-suite';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { IgxCell } from '../selection/selection.service';
 
 const CELL_CSS_CLASS = '.igx-grid__td';
 
@@ -153,8 +152,10 @@ describe('IgxGrid - CRUD operations #grid', () => {
         fix.detectChanges();
 
         const row = grid.rowList.toArray()[0];
+        // TODO: onRowEdit should emit updated rowData - issue #7304
         const args: IGridEditEventArgs = {
             rowID: 1,
+            rowData: { index: 1, value: 1 },
             oldValue: { index: 1, value: 1 },
             newValue: { index: 200, value: 200 },
             cancel: false,
@@ -174,15 +175,15 @@ describe('IgxGrid - CRUD operations #grid', () => {
         fix.detectChanges();
 
         const cell = grid.getCellByColumn(0, 'index');
-        const igxCell: IgxCell = grid.crudService.createCell(cell);
-        igxCell.editValue = 'change';
+
+        // TODO: onCellEdit should emit updated rowData - issue #7304
         const args: IGridEditEventArgs = {
             rowID: cell.cellID.rowID,
             cellID: cell.cellID,
+            rowData: { index: 1, value: 1 },
             oldValue: 1,
             newValue: 200,
             cancel: false,
-            cell: igxCell,
             column: cell.column,
             owner: grid
         };