+++ projects/igniteui-angular/src/lib/grids/tree-grid/tree-grid-crud.spec.ts
@@ -11,7 +11,6 @@ import { configureTestSuite } from '../../test-utils/configure-suite';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { GridFunctions } from '../../test-utils/grid-functions.spec';
 import { DebugElement } from '@angular/core';
-import { IgxCell } from '../selection/selection.service';
 
 const CELL_CSS_CLASS = '.igx-grid__td';
 
@@ -368,8 +367,10 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 fix.detectChanges();
                 tick(16);
                 const rowComponent = treeGrid.getRowByKey(999);
+                // TODO: onRowEdit should emit updated rowData - issue #7304
                 expect(treeGrid.onRowEdit.emit).toHaveBeenCalledWith({
                     rowID: 147,
+                    rowData: oldRow,
                     oldValue: oldRow,
                     newValue: newRow,
                     cancel: false,
@@ -397,8 +398,10 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 treeGrid.updateRow(newRow, 299);
                 fix.detectChanges();
 
+                // TODO: onRowEdit should emit updated rowData - issue #7304
                 expect(treeGrid.onRowEdit.emit).toHaveBeenCalledWith({
                     rowID: 299,
+                    rowData: oldRow,
                     oldValue: oldRow,
                     newValue: newRow,
                     cancel: false,
@@ -426,8 +429,10 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 treeGrid.getRowByKey(299).update(newRow);
                 fix.detectChanges();
 
+                // TODO: onRowEdit should emit updated rowData - issue #7304
                 expect(treeGrid.onRowEdit.emit).toHaveBeenCalledWith({
                     rowID: 299,
+                    rowData: oldRow,
                     oldValue: oldRow,
                     newValue: newRow,
                     cancel: false,
@@ -449,22 +454,22 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 verifyCellValue(fix, 6, 'Age', '25');
                 verifyRowsCount(fix, 3, 10);
 
+                const cellComponent = treeGrid.getCellByKey(299, 'Age');
+                const initialRowData = {...cellComponent.rowData};
                 // Update cell on level 3
                 const oldCellValue = treeGrid.getCellByKey(299, 'Age').value;
                 const newCellValue = 18;
                 treeGrid.updateCell(newCellValue, 299, 'Age');
                 fix.detectChanges();
 
-                const cellComponent = treeGrid.getCellByKey(299, 'Age');
-                const igxCell: IgxCell = treeGrid.crudService.createCell(cellComponent);
-                igxCell.value = oldCellValue;
+                // TODO: onCellEdit should emit updated rowData - issue #7304
                 expect(treeGrid.onCellEdit.emit).toHaveBeenCalledWith({
                     rowID: cellComponent.cellID.rowID,
                     cellID: cellComponent.cellID,
+                    rowData: initialRowData,
                     oldValue: oldCellValue,
                     newValue: newCellValue,
                     cancel: false,
-                    cell: igxCell,
                     column: cellComponent.column,
                     owner: treeGrid
                 });
@@ -484,22 +489,23 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 verifyCellValue(fix, 6, 'Age', '25');
                 verifyRowsCount(fix, 3, 10);
 
+                const cellComponent = treeGrid.getCellByKey(299, 'Age');
+                const initialRowData = {...cellComponent.rowData};
+
                 // Update cell on level 3
                 const oldCellValue = treeGrid.getCellByKey(299, 'Age').value;
                 const newCellValue = 18;
                 treeGrid.getCellByKey(299, 'Age').update(newCellValue);
                 fix.detectChanges();
 
-                const cellComponent = treeGrid.getCellByKey(299, 'Age');
-                const igxCell: IgxCell = treeGrid.crudService.createCell(cellComponent);
-                igxCell.value = oldCellValue;
+                // TODO: onCellEdit should emit updated rowData - issue #7304
                 expect(treeGrid.onCellEdit.emit).toHaveBeenCalledWith({
                     rowID: cellComponent.cellID.rowID,
                     cellID: cellComponent.cellID,
+                    rowData: initialRowData,
                     oldValue: oldCellValue,
                     newValue: newCellValue,
                     cancel: false,
-                    cell: igxCell,
                     column: cellComponent.column,
                     owner: treeGrid
                 });
@@ -537,8 +543,10 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 fix.detectChanges();
 
                 const rowComponent = treeGrid.getRowByKey(1);
+                // TODO: onRowEdit should emit updated rowData - issue #7304
                 expect(treeGrid.onRowEdit.emit).toHaveBeenCalledWith({
                     rowID: 1,
+                    rowData: oldRow,
                     oldValue: oldRow,
                     newValue: newRow,
                     cancel: false,
@@ -570,8 +578,10 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 fix.detectChanges();
 
                 const rowComponent = treeGrid.getRowByKey(999);
+                // TODO: onRowEdit should emit updated rowData - issue #7304
                 expect(treeGrid.onRowEdit.emit).toHaveBeenCalledWith({
                     rowID: 1,
+                    rowData: oldRow,
                     oldValue: oldRow,
                     newValue: newRow,
                     cancel: false,
@@ -602,8 +612,10 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 fix.detectChanges();
 
                 const rowComponent = treeGrid.getRowByKey(888);
+                // TODO: onRowEdit should emit updated rowData - issue #7304
                 expect(treeGrid.onRowEdit.emit).toHaveBeenCalledWith({
                     rowID: 7,
+                    rowData: oldRow,
                     oldValue: oldRow,
                     newValue: newRow,
                     cancel: false,
@@ -631,8 +643,10 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 treeGrid.getRowByKey(7).update(newRow);
                 fix.detectChanges();
 
+                // TODO: onRowEdit should emit updated rowData - issue #7304
                 expect(treeGrid.onRowEdit.emit).toHaveBeenCalledWith({
                     rowID: 7,
+                    rowData: oldRow,
                     oldValue: oldRow,
                     newValue: newRow,
                     cancel: false,
@@ -663,8 +677,10 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 fix.detectChanges();
 
                 const rowComponent = treeGrid.getRowByKey(4); // original component: Name = 'Jack Simon'
+                // TODO: onRowEdit should emit updated rowData - issue #7304
                 expect(treeGrid.onRowEdit.emit).toHaveBeenCalledWith({
                     rowID: 7,
+                    rowData: oldRow,
                     oldValue: oldRow,
                     newValue: newRow,
                     cancel: false,
@@ -688,22 +704,23 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 verifyCellValue(fix, 3, 'Name', 'Debra Morton');
                 verifyRowsCount(fix, 8, 8);
 
+                const cellComponent = treeGrid.getCellByKey(7, 'Name');
+                const initialRowData = {...cellComponent.rowData};
+
                 // Update cell on level 3
                 const oldCellValue = treeGrid.getCellByKey(7, 'Name').value;
                 const newCellValue = 'Michael Myers';
                 treeGrid.updateCell(newCellValue, 7, 'Name');
                 fix.detectChanges();
 
-                const cellComponent = treeGrid.getCellByKey(7, 'Name');
-                const igxCell: IgxCell = treeGrid.crudService.createCell(cellComponent);
-                igxCell.value = oldCellValue;
+                // TODO: onCellEdit should emit updated rowData - issue #7304
                 expect(treeGrid.onCellEdit.emit).toHaveBeenCalledWith({
                     rowID: cellComponent.cellID.rowID,
                     cellID: cellComponent.cellID,
+                    rowData: initialRowData,
                     oldValue: oldCellValue,
                     newValue: newCellValue,
                     cancel: false,
-                    cell: igxCell,
                     column: cellComponent.column,
                     owner: treeGrid
                 });
@@ -723,6 +740,9 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 verifyCellValue(fix, 3, 'Name', 'Debra Morton');
                 verifyRowsCount(fix, 8, 8);
 
+                const cellComponent = treeGrid.getCellByKey(7, 'Name');
+                const initialRowData = {...cellComponent.rowData};
+
                 // Update cell on level 3
                 const oldCellValue = treeGrid.getCellByKey(7, 'Name').value;
                 const newCellValue = 'Michael Myers';
@@ -730,16 +750,14 @@ describe('IgxTreeGrid - CRUD #tGrid', () => {
                 treeGrid.getCellByKey(7, 'Name').update(newCellValue);
                 fix.detectChanges();
 
-                const cellComponent = treeGrid.getCellByKey(7, 'Name');
-                const igxCell: IgxCell = treeGrid.crudService.createCell(cellComponent);
-                igxCell.value = oldCellValue;
+                // TODO: onCellEdit should emit updated rowData - issue #7304
                 expect(treeGrid.onCellEdit.emit).toHaveBeenCalledWith({
                     rowID: cellComponent.cellID.rowID,
                     cellID: cellComponent.cellID,
+                    rowData: initialRowData,
                     oldValue: oldCellValue,
                     newValue: newCellValue,
                     cancel: false,
-                    cell: igxCell,
                     column: cellComponent.column,
                     owner: treeGrid
                 });