+++ CHANGELOG.md
@@ -6,7 +6,8 @@ All notable changes for each version of this project will be documented in this 
 
 ### New Features
 - `IgxGrid`, `IgxTreeGrid`, `IgxHierarchicalGrid`
-    - The event arguments of `onCellEdit`, `onCellEditEnter` and `onCellEditCancel` events will contain a reference to the edited cell, as well as a reference to the column where the edited cell is.
+    - The event arguments of `onCellEdit`, `onCellEditEnter` and `onCellEditCancel` events will contain a reference to the row data, as well as a reference to the column.
+    - The event arguments of `onRowEdit`, `onRowEditEnter` and `onRowEditCancel` events will contain a reference to the row data.
 
 ### General
 - `igx-select`, `igx-combo`, `igx-drop-down`